requirements.txt
Django==4.0.5
django-cors-headers==3.13.0
django-filter==21.1
djangorestframework==3.13.1
mysqlclient==2.1.1
Pillow==9.0.1
